Static lib and less warnings

+++ L3gion/src/L3gion/KeyCodes.cpp
@@ -8,7 +8,7 @@ namespace L3gion
 		auto it = std::find(glfwKeysBuffer.begin(), glfwKeysBuffer.end(), keycode);
 
 		if (it != glfwKeysBuffer.end())
-			return std::distance(glfwKeysBuffer.begin(), it);
+			return (int)std::distance(glfwKeysBuffer.begin(), it);
 
 		LG_CORE_ERROR("The GLFW_KEY provided ({0}) does not match the LG_KEYs!!", keycode);
 
@@ -20,7 +20,7 @@ namespace L3gion
 		auto it = std::find(imGuiKeysBuffer.begin(), imGuiKeysBuffer.end(), keycode);
 
 		if (it != imGuiKeysBuffer.end())
-			return std::distance(imGuiKeysBuffer.begin(), it);
+			return (int)std::distance(imGuiKeysBuffer.begin(), it);
 
 		LG_CORE_ERROR("The IMGUI_KEY provided ({0}) does not match the LG_KEYs!!", keycode);
